requirements.txt
gallery-dl>=1.26.0
rich>=13.0.0
typer>=0.9.0
PyQt6>=6.5.0
pyyaml>=6.0
requests>=2.31.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0